fixes bug in tabs component
