--- packages/ripple/package.json
@@ -1,26 +1,0 @@
-
-  {
-    "name": "@leafygreen-ui/ripple",
-    "version": "0.9.0",
-    "description": "leafyGreen UI Kit Ripple",
-    "main": "./dist/index.js",
-    "module": "./dist/esm/index.js",
-    "types": "./dist/index.d.ts",
-    "typesVersions": {
-      "<3.9": { "*": ["ts3.4/*"] }
-    },
-    "scripts": {
-      "build": "../../node_modules/.bin/rollup --config ../../rollup.config.js"
-    },
-    "license": "Apache-2.0",
-    "publishConfig": {
-      "access": "public"
-    },
-    "dependencies": {
-      "polished": "^4.1.1",
-      "@leafygreen-ui/palette": "^3.1.1"
-    },
-    "devDependencies": {
-      "@leafygreen-ui/lib": "^6.1.0"
-    }
-  }
