--- packages/ripple/tsconfig.json
@@ -1,17 +1,0 @@
-{
-  "extends": "../../package.tsconfig.json",
-  "compilerOptions": {
-    "declarationDir": "dist",
-    "outDir": "dist",
-    "rootDir": "src"
-  },
-  "include": ["src/**/*"],
-  "references": [
-    {
-      "path": "../lib"
-    },
-    {
-      "path": "../palette"
-    }
-  ]
-}
